bpf: extract L4 port+checksum rewrite helper for sharing
Extract the port+checksum rewrite logic out of
snat_v4/v6_rewrite_headers() into l4_rewrite_port_and_csum(), on top
of the L3 helper extraction, so the load balancer's upcoming
RevDNAT-of-ICMP-Error work can reuse it instead of duplicating it.
No functional change.
